Require security token from GitHub
This commit checks for a secret key only I know in the request header. Since I
have not put the key in the GH webhook configuration yet, when I push this
commit the server I have running against this branch will reject the HTTP
ppost request sent by GH. When I add the token and resend the same request, it
will be accepted.